+++ product-service/src/test/java/com/msa/service/product/eventUseCase/ConsumeOrderCompletedEvent.java
@@ -1,8 +1,11 @@
 package com.msa.service.product.eventUseCase;
 
+import com.msa.application.ProductFacade;
 import com.msa.application.events.EventUseCase;
 import com.msa.domain.Product;
 import com.msa.domain.repository.ProductRepository;
+import com.msa.domain.service.CategoryService;
+import com.msa.domain.service.ProductCategoryService;
 import com.msa.domain.service.ProductService;
 import com.msa.domain.vo.ProductInfo;
 import com.msa.infrastructure.kafka.Events;
@@ -24,6 +27,12 @@ public class ConsumeOrderCompletedEvent {
     ProductRepository productRepository;
     @Autowired
     EventUseCase eventUseCase;
+    @Autowired
+    ProductFacade facade;
+    @Autowired
+    CategoryService categoryService;
+    @Autowired
+    ProductCategoryService productCategoryService;
 
 
     @Test
@@ -36,7 +45,7 @@ public class ConsumeOrderCompletedEvent {
 
 
         //when
-        eventUseCase.consumeOrderCompletedEvent(event);
+        facade.consumeOrderCompletedEvent(event);
 
         //then
         Product targetProduct = productRepository.findById(saved.getId());
@@ -61,13 +70,12 @@ public class ConsumeOrderCompletedEvent {
         for (int i = 0; i < 4; i++) {
             executorService.execute(() -> {
                 try {
-                    eventUseCase.consumeOrderCompletedEvent(event);
+                    facade.consumeOrderCompletedEvent(event);
                 } catch (ObjectOptimisticLockingFailureException e) {
                     System.out.println("동시 접근 발생!");
                 }
             });
         }
-
 
         Thread.sleep(3000);
         Product targetProduct = productRepository.findById(saved.getId());
@@ -93,7 +101,7 @@ public class ConsumeOrderCompletedEvent {
         for (int i = 0; i < 4; i++) {
             executorService.execute(() -> {
                 try {
-                    eventUseCase.consumeOrderCompletedEvent(event);
+                    facade.consumeOrderCompletedEvent(event);
                 } catch (ObjectOptimisticLockingFailureException e) {
                     System.out.println("동시 접근 발생!");
                 }
