+++ product-service/src/test/resources/application.yml
@@ -13,3 +13,4 @@ spring:
     properties:
       hibernate:
         dialect: org.hibernate.dialect.H2Dialect
+    show-sql: true
